Improve function call by making less parameters being obligatory
